refactoring code + improving exit handler when exit is/isn't asked by user expressively
