+++ src/main/java/main/process/InteraactionSceneNamedProcessCreator.java
@@ -1,0 +1,32 @@
+package main.process;
+
+import javafx.application.Platform;
+import main.UI.menu.GraphicalMenus;
+import main.process.xdotoolProcess.GoogleChromeXdotoolProcessCreator;
+import main.utils.NamedProcess;
+import main.utils.UtilsOS;
+
+import java.io.IOException;
+
+public class InteraactionSceneNamedProcessCreator implements AppNamedProcessCreator {
+
+    ProcessBuilder processBuilder;
+
+    @Override
+    public void setUpProcessBuilder() {
+        processBuilder = new ProcessBuilder(AppNamedProcessCreator.getBrowser(),
+                "--kiosk",
+                "--window-position=0,0",
+                "--disable-gpu",
+                "--no-sandbox",
+                "--fullscreen",
+                "http://interaactionScene.net/");
+    }
+
+    @Override
+    public NamedProcess start(GraphicalMenus graphicalMenus) {
+        return AppNamedProcessCreator.createProcress(new GoogleChromeXdotoolProcessCreator(),processBuilder,graphicalMenus,"InteraactionScene");
+
+    }
+
+}
