+++ src/main/java/main/process/WifiNamedProcessCreator.java
@@ -1,0 +1,27 @@
+package main.process;
+
+import javafx.application.Platform;
+import main.UI.menu.GraphicalMenus;
+import main.process.xdotoolProcess.GoogleChromeXdotoolProcessCreator;
+import main.process.xdotoolProcess.WifiXdotoolProcessCreator;
+import main.utils.NamedProcess;
+import main.utils.UtilsOS;
+
+import java.io.IOException;
+
+public class WifiNamedProcessCreator implements AppNamedProcessCreator {
+
+    ProcessBuilder processBuilder;
+
+    @Override
+    public void setUpProcessBuilder() {
+        processBuilder = new ProcessBuilder("nm-connection-editor");
+    }
+
+    @Override
+    public NamedProcess start(GraphicalMenus graphicalMenus) {
+        return AppNamedProcessCreator.createProcress(new WifiXdotoolProcessCreator(),processBuilder,graphicalMenus,"Network Manager");
+
+    }
+
+}
