--- src/main/java/main/process/WifiProcess.java
@@ -1,51 +1,0 @@
-package main.process;
-
-import javafx.application.Platform;
-import main.UI.menu.GraphicalMenus;
-import main.process.xdotoolProcess.WifiXdotoolProcess;
-import main.utils.NamedProcess;
-
-import java.io.IOException;
-
-public class WifiProcess implements AppProcess {
-
-    ProcessBuilder processBuilder;
-
-    @Override
-    public void setUpProcessBuilder() {
-        processBuilder = new ProcessBuilder("nm-connection-editor");
-    }
-
-    @Override
-    public NamedProcess start(GraphicalMenus graphicalMenus) {
-        try {
-            NamedProcess namedProcess = new NamedProcess();
-            WifiXdotoolProcess gcxp = new WifiXdotoolProcess();
-            gcxp.setUpProcessBuilder();
-            gcxp.start();
-
-            Process process = processBuilder.inheritIO().start();
-
-            process.onExit().thenRun(
-                    new Runnable() {
-                        @Override
-                        public void run() {
-                            Platform.runLater((() -> {
-                                graphicalMenus.getHomeScreen().showCloseMenuIfProcessNotNull();
-                                graphicalMenus.primaryStage.show();
-                                graphicalMenus.primaryStage.toFront();
-                            }));
-                        }
-                    }
-            );
-            AppProcess.startWindowIdSearcher(graphicalMenus, "wifi");
-            namedProcess.set(process);
-            namedProcess.setName("Network Manager");
-            return namedProcess;
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-}
