+++ src/main/java/main/process/YoutubeNamedProcessCreator.java
@@ -1,0 +1,31 @@
+package main.process;
+
+import javafx.application.Platform;
+import main.UI.menu.GraphicalMenus;
+import main.process.xdotoolProcess.GoogleChromeXdotoolProcessCreator;
+import main.utils.NamedProcess;
+import main.utils.UtilsOS;
+
+import java.io.IOException;
+
+public class YoutubeNamedProcessCreator implements AppNamedProcessCreator {
+
+    ProcessBuilder processBuilder;
+
+    @Override
+    public void setUpProcessBuilder() {
+        processBuilder = new ProcessBuilder(AppNamedProcessCreator.getBrowser(),
+                "--kiosk",
+                "--window-position=0,0",
+                "--disable-gpu",
+                "--no-sandbox",
+                "--fullscreen",
+                "https://www.youtube.com");
+    }
+
+    @Override
+    public NamedProcess start(GraphicalMenus graphicalMenus) {
+        return AppNamedProcessCreator.createProcress(new GoogleChromeXdotoolProcessCreator(),processBuilder,graphicalMenus,"YouTube");
+
+    }
+}
